+++ libs/common/src/lib/interfaces/user-settings.interface.ts
@@ -7,6 +7,7 @@ export interface UserSettings {
   colorScheme?: ColorScheme;
   dateRange?: DateRange;
   emergencyFund?: number;
+  'filters.tags'?: string[];
   isExperimentalFeatures?: boolean;
   isRestrictedView?: boolean;
   language?: string;